Change content-type to x-ndjson, sent by searchkit to the search engine.
Searchkit was sending "application/json" to the search engine API,
this is technically incorrect, since the body of the request is a
list of jsons concatenated by newline. This breaks the parsing of the
request body on some (proxy) servers, for example Ruby on Rails, since
when it tries to parse a json, if finds it is broken.

This fixes #1360.